# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Knowledge Base API — used by the web app and AI agents
 * OpenAPI spec version: 0.1.0
 */
import type { UserUpdateRole } from './userUpdateRole';

export interface UserUpdate {
  email?: string;
  /** @minLength 1 */
  name?: string;
  role?: UserUpdateRole;
  /** @minLength 8 */
  password?: string;
}